Replace non-standard ssize_t.

+++ code_examples/algorithms/for_each_n_code.h
@@ -9,7 +9,7 @@ std::for_each_n(std::execution::par_unseq,
 auto it = final_ranking.begin();
 uint32_t page = 0;
 while (it != final_ranking.end()) {
-	ssize_t cnt = std::min(PAGE_SIZE, final_ranking.end() - it);
+	size_t cnt = std::min(PAGE_SIZE, size_t(final_ranking.end() - it));
 	std::for_each_n(it, cnt, [page](const Player& p) {
 		store_final_score(page, p.display_name, p.score);
 	});
